Remediate ThinkingBox client and evaluator boundaries
