+++ .github/workflows/test.yml
@@ -98,6 +98,10 @@ jobs:
 
   test-thinkingbox-env:
     runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.11", "3.12"]
 
     steps:
       - name: Checkout code
@@ -108,16 +112,19 @@ jobs:
         with:
           version: ${{ env.UV_VERSION }}
 
-      - name: Set up Python 3.12
-        run: uv python install 3.12
+      - name: Set up Python ${{ matrix.python-version }}
+        run: uv python install ${{ matrix.python-version }}
 
       - name: Install ThinkingBox environment dependencies
         working-directory: envs/thinkingbox_env
-        run: uv sync --python 3.12 --frozen --all-groups --all-extras
+        run: |
+          uv sync --python ${{ matrix.python-version }} \
+            --frozen --all-groups --all-extras
 
       - name: Run ThinkingBox environment tests
         run: |
-          uv run --python 3.12 --frozen --project envs/thinkingbox_env \
+          uv run --python ${{ matrix.python-version }} \
+            --frozen --project envs/thinkingbox_env \
             --all-groups --all-extras \
             pytest tests/envs/test_thinkingbox_env.py -v --tb=short
         env:
