+++ envs/thinkingbox_env/.dockerignore
@@ -1,6 +1,9 @@
 .venv
 __pycache__
 .pytest_cache
+build
+dist
+*.egg-info
 *.pyc
 *.pyo
 logs
