+++ envs/thinkingbox_env/__init__.py
@@ -13,6 +13,7 @@ from .models import (
     SubmitMessageAction,
     SubmittedToolCall,
     ThinkingBoxAction,
+    ThinkingBoxExecutionProvenance,
     ThinkingBoxObservation,
     ThinkingBoxState,
     ToolCallResult,
@@ -25,6 +26,7 @@ __all__ = [
     "SubmittedToolCall",
     "ThinkingBoxAction",
     "ThinkingBoxEnv",
+    "ThinkingBoxExecutionProvenance",
     "ThinkingBoxObservation",
     "ThinkingBoxState",
     "ToolCallResult",
